requirements.txt
streamlit==1.45.0
mysql-connector-python==9.5.0
bcrypt==4.3.0
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
streamlit-js-eval==0.1.7
Pillow==11.2.1
requests==2.32.3
